refactor(db): extract terminal_tab into db/terminal.rs
