+++ src/db/terminal.rs
@@ -1,0 +1,175 @@
+//! Terminal tab CRUD methods on `Database`.
+//!
+//! This file contributes a `impl Database { ... }` block to the type defined
+//! in `super::Database`. Multiple `impl` blocks on the same type across files
+//! are idiomatic Rust; the public method paths resolve identically to a
+//! single-block layout.
+
+use rusqlite::params;
+
+use crate::model::TerminalTab;
+
+use super::Database;
+
+impl Database {
+    // --- Terminal Tabs ---
+
+    pub fn insert_terminal_tab(&self, tab: &TerminalTab) -> Result<(), rusqlite::Error> {
+        self.conn.execute(
+            "INSERT INTO terminal_tabs (id, workspace_id, title, is_script_output, sort_order)
+             VALUES (?1, ?2, ?3, ?4, ?5)",
+            params![
+                tab.id,
+                tab.workspace_id,
+                tab.title,
+                tab.is_script_output as i32,
+                tab.sort_order,
+            ],
+        )?;
+        Ok(())
+    }
+
+    pub fn max_terminal_tab_id(&self) -> Result<i64, rusqlite::Error> {
+        self.conn.query_row(
+            "SELECT COALESCE(MAX(id), 0) FROM terminal_tabs",
+            [],
+            |row| row.get(0),
+        )
+    }
+
+    pub fn list_terminal_tabs_by_workspace(
+        &self,
+        workspace_id: &str,
+    ) -> Result<Vec<TerminalTab>, rusqlite::Error> {
+        let mut stmt = self.conn.prepare(
+            "SELECT id, workspace_id, title, is_script_output, sort_order, created_at
+             FROM terminal_tabs WHERE workspace_id = ?1 ORDER BY sort_order, id",
+        )?;
+        let rows = stmt.query_map(params![workspace_id], |row| {
+            let is_script: i32 = row.get(3)?;
+            Ok(TerminalTab {
+                id: row.get(0)?,
+                workspace_id: row.get(1)?,
+                title: row.get(2)?,
+                is_script_output: is_script != 0,
+                sort_order: row.get(4)?,
+                created_at: row.get(5)?,
+            })
+        })?;
+        rows.collect()
+    }
+
+    pub fn delete_terminal_tab(&self, id: i64) -> Result<(), rusqlite::Error> {
+        self.conn
+            .execute("DELETE FROM terminal_tabs WHERE id = ?1", params![id])?;
+        Ok(())
+    }
+
+    #[allow(dead_code)]
+    pub fn delete_terminal_tabs_for_workspace(
+        &self,
+        workspace_id: &str,
+    ) -> Result<(), rusqlite::Error> {
+        self.conn.execute(
+            "DELETE FROM terminal_tabs WHERE workspace_id = ?1",
+            params![workspace_id],
+        )?;
+        Ok(())
+    }
+
+    #[allow(dead_code)]
+    pub fn update_terminal_tab_title(&self, id: i64, title: &str) -> Result<(), rusqlite::Error> {
+        self.conn.execute(
+            "UPDATE terminal_tabs SET title = ?1 WHERE id = ?2",
+            params![title, id],
+        )?;
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::db::test_support::*;
+
+    fn make_terminal_tab(id: i64, ws_id: &str, title: &str) -> TerminalTab {
+        TerminalTab {
+            id,
+            workspace_id: ws_id.into(),
+            title: title.into(),
+            is_script_output: false,
+            sort_order: 0,
+            created_at: String::new(),
+        }
+    }
+
+    #[test]
+    fn test_insert_and_list_terminal_tabs() {
+        let db = setup_db_with_workspace();
+        db.insert_terminal_tab(&make_terminal_tab(1, "w1", "Terminal 1"))
+            .unwrap();
+        db.insert_terminal_tab(&make_terminal_tab(2, "w1", "Terminal 2"))
+            .unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert_eq!(tabs.len(), 2);
+        assert_eq!(tabs[0].title, "Terminal 1");
+        assert_eq!(tabs[1].title, "Terminal 2");
+    }
+
+    #[test]
+    fn test_terminal_tabs_filtered_by_workspace() {
+        let db = setup_db_with_workspace();
+        db.insert_workspace(&make_workspace("w2", "r1", "feature"))
+            .unwrap();
+        db.insert_terminal_tab(&make_terminal_tab(1, "w1", "T1"))
+            .unwrap();
+        db.insert_terminal_tab(&make_terminal_tab(2, "w2", "T2"))
+            .unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert_eq!(tabs.len(), 1);
+        assert_eq!(tabs[0].title, "T1");
+    }
+
+    #[test]
+    fn test_delete_terminal_tab() {
+        let db = setup_db_with_workspace();
+        db.insert_terminal_tab(&make_terminal_tab(1, "w1", "Terminal 1"))
+            .unwrap();
+        db.delete_terminal_tab(1).unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert!(tabs.is_empty());
+    }
+
+    #[test]
+    fn test_terminal_tabs_cascade_on_workspace_delete() {
+        let db = setup_db_with_workspace();
+        db.insert_terminal_tab(&make_terminal_tab(1, "w1", "Terminal 1"))
+            .unwrap();
+        db.insert_terminal_tab(&make_terminal_tab(2, "w1", "Terminal 2"))
+            .unwrap();
+        db.delete_workspace("w1").unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert!(tabs.is_empty());
+    }
+
+    #[test]
+    fn test_terminal_tab_script_output_flag() {
+        let db = setup_db_with_workspace();
+        let mut tab = make_terminal_tab(1, "w1", "npm run dev");
+        tab.is_script_output = true;
+        db.insert_terminal_tab(&tab).unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert!(tabs[0].is_script_output);
+    }
+
+    #[test]
+    fn test_update_terminal_tab_title() {
+        let db = setup_db_with_workspace();
+        db.insert_terminal_tab(&make_terminal_tab(1, "w1", "Terminal 1"))
+            .unwrap();
+        db.update_terminal_tab_title(1, "My Custom Terminal")
+            .unwrap();
+        let tabs = db.list_terminal_tabs_by_workspace("w1").unwrap();
+        assert_eq!(tabs[0].title, "My Custom Terminal");
+    }
+}
